feat: move RFID reader to dedicated app
