+++ accounts/tests.py
@@ -23,8 +23,6 @@ from django.core.exceptions import ValidationError
 from django.core.management import call_command
 from django.db import IntegrityError
 from .backends import LocalhostAdminBackend
-from channels.testing import WebsocketCommunicator
-from config.asgi import application
 
 
 class DefaultAdminTests(TestCase):
@@ -383,11 +381,4 @@ class RFIDFixtureTests(TestCase):
         self.assertIn(account, tag.accounts.all())
         self.assertEqual(tag.accounts.count(), 1)
 
-
-class RFIDConsumerTests(TransactionTestCase):
-    async def test_websocket_connects(self):
-        communicator = WebsocketCommunicator(application, "/ws/rfid/")
-        connected, _ = await communicator.connect()
-        self.assertTrue(connected)
-        await communicator.disconnect()
