+++ rfid/routing.py
@@ -1,5 +1,4 @@
 from django.urls import path
-
 from . import consumers
 
 websocket_urlpatterns = [
