+++ tests/test_urls_autodiscover.py
@@ -13,6 +13,11 @@ def test_accounts_url_autodiscovered():
     assert match.view_name == "rfid-login"
 
 
+
+def test_rfid_url_autodiscovered():
+    match = resolve("/rfid/")
+    assert match.view_name == "rfid-reader"
+
 def test_references_custom_prefix():
     match = resolve("/ref/")
     assert match.view_name == "references:generator"
